修正 ItemCreator damage 在 fromItem 時不會套用的問題 (ex: 噴濺藥水 make 變成喝的)

+++ src/main/java/org/mineacademy/fo/SerializeUtil.java
@@ -413,7 +413,7 @@ public final class SerializeUtil {
 		if (bukkitWorld == null)
 			throw new InvalidWorldException("Location with invalid world '" + world + "': " + raw + " (Doesn't exist)", world);
 
-		final int x = Integer.parseInt(parts[1]), y = Integer.parseInt(parts[2]), z = Integer.parseInt(parts[3]);
+		final double x = Double.parseDouble(parts[1]), y = Double.parseDouble(parts[2]), z = Double.parseDouble(parts[3]);
 		final float yaw = Float.parseFloat(parts.length == 6 ? parts[4] : "0"), pitch = Float.parseFloat(parts.length == 6 ? parts[5] : "0");
 
 		return new Location(bukkitWorld, x, y, z, yaw, pitch);
